+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 * Bug fixes
   * Fix content negotiation when media type params do not match
+  * Let content types provided correctly match all or none accept-extensions
 
 ## v0.5.0